refactor: changed app errors package name
